requirements.txt
annotated-types==0.7.0
anyio==4.7.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
psycopg2==2.9.10
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
python-multipart==0.0.19
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
